tst_QQuickTextInput: Stabilize clipboard related functions on XCB
Even if compiled with QT_CONFIG(clipboard), a functional clipboard
might not be available on XCB (e.g. due to restrictions on CI VMs).

This patch adds an explicit test for XCB to platformquirks_p.h.
It skips the test functions canPaste() and canPasteEmpty().
As a drive-by it constifies local variables in canPaste() and
middleClickPaste() and unifies the skip message.

It also removes the BLACKLIST entries for clipboard-related tests on
opensuse.

Fixes: QTBUG-95940
Pick-to: 6.5
Change-Id: If6da9b589e98c1f63435f3d444567ce310ddee8a
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>