Hook up GPU frame timing to QQuickGraphicsConfig and the logs
The idea being that setting QSG_RHI_PROFILE=1 will not just enable
debug markers (which is only actively used in Quick 3D right now),
but also timestamps, and then we can start showing GPU frame times
from the GPU side with Vulkan, Metal, and D3D11 atm.
(the only catch is that the value is heavily asynchronous and is
always referring to a previous frame, usually current - 2, but may
be up to current - 4 etc.)

In addition to the env.var., the setter/getter API is added to
QQuickGraphicsConfiguration. In the end they all maps to the
QRhi::Timestamps flag.

Once enabled, the renderloop timing log, both with threaded and
basic, will start including a new line, as long as the underlying
3D API supports timestamp queries or similar. (in practice this is
implemented for Metal, Vulkan, and D3D11 for now)

[ChangeLog][QtQuick][Scenegraph] The scenegraph render loop timing
logs can now show GPU-side times as well, as long as the application
is launched with the env.var. QSG_RHI_PROFILE=1 or the feature is
enabled via the new APIs in QQuickGraphicsConfiguration.

Change-Id: I42202882b418e12028d6a287c19352a8fb893365
Reviewed-by: Andy Nichols <andy.nichols@qt.io>