MultiTouchArea: only subscribe to hover events when enabled
On macOS we always subscribe to hover events in the constructor so
that we can enable touch event handling when the mouse enters the
area. However, a disabled item should not accept hover events, so
we also need to override itemChange so that hover-acceptance is
flipped based on the enabled state.

Items should also never accept hover events, they should always
propagate. As the events are accepted by default and ignored in the
default event handler implementation in QQuickItem, we need to
ignore explicitly when we override. Do that by calling the parent
class implementation explicitly.

A disabled MultiPointTouchArea should also not subscribe to touch
events, even when the mouse enters.

Fixes: QTBUG-107037
Pick-to: 6.5
Change-Id: Ifa57cb2af4fbf11a0c8eacde87d80e60afa3f895
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>