TextEdit: fix wrong RTL text alignment when used with a ColumnLayout
Due to some old bug in QTextDocument, invalidating document's
width (via setTextWidth(-1)) also breaks RTL text alignment.
It was a reason of broken text alignment in case when TextEdit
control is placed into a layout. To avoid this, we have to call
setTextWidth("idealWidth") all the time when an explicit width
is not set and text wrapping is not used.
As a drive-by change, we slightly polished the surrounding code
to make it a bit clear.

Fixes: QTBUG-112858
Pick-to: 6.5
Change-Id: I59cb0618d939d05fedbbabc59a8aae0e2ecabaee
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>