ComboBox: don't open on key release unless press was handled
If a Keys handler eats the press of the activation key, then the release
event shouldn't show the popup. The default keyPressEvent handler sets
the pressed state to true, so we can test that state to see if we should
act on the corresponding release. The key release event should still be
accepted to stop propagation.

Add a test case.

Pick-to: 6.5 6.2
Fixes: QTBUG-109721
Change-Id: Icb76453733d05cba59fdfb365b0cebf710be5f01
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>