StackView: emit depthChanged when clearing with transition
When a transition is used in the call to clear(), then we pop the top
element off the stack and start the transition for it. If we read the
old depth after doing that, then it will be off by one and prevent the
emission of the changed signal if that was the last element on the
stack.

So read the old depth already before we pop, and add a test.

Fixes: QTBUG-84920
Pick-to: 6.5 6.2
Change-Id: Idae619efa25729fafbf238cb4db62472042b1cb7
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>