Overlay: clean up grabbers on touch release
The overlay installs itself as an event filter on the QQuickWindow,
passes the event on to the delivery agent directly, and eats the event.
This bypasses regular event processing in QQuickWindow::Event, including
the clearing of any grabs when an event point is released. This leaves
handlers in an inconsistent state.

To fix this, factor the clean-up routine out into a separate function
that the overlay can call at the end of processing.

Add a test project for QQuickOverlay. Not making it a QML test as there
are no qml files yet, and no need to unit-test this behavior with all
styles.

Pick-to: 6.5 6.6
Fixes: QTBUG-110328
Change-Id: Iebe7a252de9919d4958dcda405d24e716353764a
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>