doc: Expand module docs for Qt Quick Shapes
Give an overview of the module and some handy links and tips. This
also adds a note on how to get anti-aliasing with the shapes, as well
as a link to PathText as one of the supported element types.

Fixes: QTBUG-106529
Task-number: QTBUG-112463
Change-Id: I137d25e2ca138454f913e92158d9286977457227
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit ee017bd9cfe852883418ed08d38b4d001c9dbc17)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>