CMake: Reference all qml resources in static qml plugin constructor
Previously, when a static library was made into a qml module, we added
resources containing the qml files, qmldir, etc, into separate object
libraries rather than directly into the static library.

We did this to ensure that the linker doesn't discard these resources
if they are not explicitly initialized with Q_INIT_RESOURCE in user
code.

This proves to be a problem for projects that link directly to the
static libraries instead of using target names, because the object
libraries that contain the resources are never referenced in the
cmake project and thus not linked in.
This tends to happen in projects that ship their own static qml
modules without accompanying target information.

To improve the situation, move all the resources back into the
backing libraries, and reference the resources inside the static
qml plugin constructor.

This way a project that wishes to link to libraries directly can do
the following and expect a working project:
- link to the static backing library
- link to the static qml plugin
- add Q_IMPORT_QML_PLUGIN(PluginName) in their project code.

This will ensure all resources are kept and initialized.

To implement that, use the new macros introduced in QtCore's
qtsymbolmacros.h which generate code to ensure symbols are not
discarded.

It's based on the same idea we use to keep the type registration
function: a volatile reference to the symbol to keep.

Add a test that confirms all this works.

Fixes: QTBUG-110243
Change-Id: I00a127ef17bc852e360fb7a2dc58ee20522424cd
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>