StyleGenerator: support writing regexp when seatching for Figma atoms
When searching for an atom in Figma with a given name, use a
regexp. This will allow us to write names in the config file
that can match e.g the beginning of a name, handle spaces, or
ignore any extra text (or spaces) appended to the name. It can
e.g happen that the name has some info in paranthesis added
behind the name.

Task-number: QTBUG-111560
Change-Id: I54c41bd67c0ff29fe849e169bcc5da6664e13878
Reviewed-by: Doris Verria <doris.verria@qt.io>