Fix StackLayout to keep the current visible item after insert/removal
If an item is inserted or deleted at an index less than or equal to the
current index, StackLayout.currentIndex will be updated in order to
reflect which is the current visible item. This is consistent with
QStackedLayout, QStackedWidget and TabBar

[ChangeLog][QtQuick][Layouts] StackLayout will now update currentIndex
if an Item is inserted or removed at an index less than or equal to the
current index.

Task-number: QTBUG-112691
Change-Id: Id1d687e309c8f199a4f68698b53ca7c169f17dcd
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit adc3d8b2e40ca2853b4baf5ea893021a799c21f4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>