MenuBar: Clean up and stabilize test case
The test case fails frequently and flakily locally with the Material
style, because the exit transition of the static menu interferes with
the opening of the dynamic menu. Wait for the static menu to be
completely done with the exit transition before continuing.

As a drive-by, replace all QCOMPARE(actual, true/false) patterns with
QVERIFY.

Task-number: QTBUG-104323
Pick-to: 6.5
Change-Id: Ide90f0094b05d628a0eb047248a72c35bcb7cdc3
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>