Material: fix clipped floating placeholder text
The changes to placeholder text in
20e3d1b522d1b79239e9ac4a6af47ce3648512bd and
5704741a4073ac131782c3dd73cac5cda6800a28 result in floating placeholder
text (i.e. the text shown at the top of the control when it has active
focus) being clipped when e.g. in a ScrollView.

This is probably only an issue for TextArea in practice, since you
wouldn't usually put a TextField in a ScrollView, but you can still run
into it if you clip it.

We don't want to unconditionally set topInset by default, because, as
mentioned above, these controls are not always clipped. In most cases
they are used on their own, and this issue won't affect them.
Unconditionally setting topInset would ruin the layout of existing UIs.
So, we set topInset only if the control itself clips (or its Flickable
parent in the case of TextArea).

[ChangeLog][Controls][Material] The outlined TextArea now sets topInset
by default if it or its Flickable parent clips. This avoids the
floating placeholder being clipped in those cases. The outlined
TextField sets topInset by default only if the TextField itself clips.

Fixes: QTBUG-113321
Pick-to: 6.5
Change-Id: I8555e4fc0c7a9800f76b54a84d94f4d04691bc23
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>