Clear out some locale-dependencies from XHR unit test
None of these tests care about Accept-Language, which
QHttpNetworkConnection sets automatically based on the system locale,
so testing for it (by including a line for it in .expect files),
setting it and skipping tests due to its "locale-dependence" were all
unnecessary. Just tell the test-server to {{Ignore}} the value of the
header.

This failed on Android due to one of the data files being unavailable,
so make access to that a precondition of its test.

Change-Id: I560ead8b3ace6a5df173d96aea10f8dac3ed1124
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit d0a2f511f3fd8a736ddff74f984585b8909d1d8d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>