Models: Fix condition for context object
As shown in the ComboBox test, it is possible for a context object to
coexist with required properties. We have to check if the context object
is the same as the model object.

Amends commit e9f650cad82e7eff2234cdf2d37bb9b1bb25d27d

Task-number: QTBUG-111176
Task-number: QTBUG-110980
Task-number: QTBUG-104752
Change-Id: Idfee277b5c3b78c514b2d239d73fcb6809577678
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>