qmlpreview: fix crash on exit
saveWindowPosition() uses QSettings which uses
the special file engine, which does not exist
at destructor time anymore.
This resulted in a crash.
QQmlPreviewPosition idea is that saveWindowPosition()
is called when there was a move event.
Also saving it at the constructor was just a nice to
have.

Change-Id: I650176691acf8ad1ef9ae46b76f61dafee82a56d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit db01d4ef5a250433deb99eabf23eb29437e02905)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>