StyleGenerator: implement support for 'mirrored' controls
In order to know if a control in Figma is mirrored or
not, we need to know which atoms are a part of the
contentItem, and their expected left-to-right order.
This patch will therefore add a 'contents' field to
the config file that lets you list those atoms, and
their expected order. This will then be read by the
generator tool in order to write a 'mirrored' field
into the output config.

If the style is using SPACE_BETWEEN (labeled "auto"
in Figma) as horizontal layout spacing, then we cannot
currently support that, since our controls are designed
to have fixed spacing between the atoms.

So as a fallback when this is the case, we look at the
geometries of the content item atoms, and calculate the
fixed spacing between them. The difference will be
visible if you later scale the control; In QML, the
atoms will continue to have a fixed space between them.
In Figma, the atoms would instead spread out more evenly.

Task-number: QTBUG-111560
Change-Id: I750d39de7e933ca99f77dad357d82fa0726421dd
Reviewed-by: Doris Verria <doris.verria@qt.io>