FileLocations::Tree: rename treePtr and fileLocationsPtr
Rename and add documentation to FileLocations::Tree::treePtr and
FileLocations::Tree::fileLocationsPtr.

Also remove a helper method in QQmlLSUtils that did exactly the same as
treePtr.

Both static methods allow to obtain information about the location of
some DomItem in the source code.

treePtr() is now called treeOf() and fileLocationsPtr() is
fileLocationsOf().

Change-Id: Ia6c92f7576b36727ce2fb72b729a4730b5250c3d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>