QQuickSinglePointHandler: reset pointinfo on a stray TouchBegin
Extremely fast finger drumming on a touch screen with e.g. a tap
handler can result in one event being inserted between two others.
Such unwanted events are caught in
QQuickSinglePointHandler::wantsPointerEvent() and results in a warning.

However, once the sequence of TouchBegin, TouchUpdate, TouchEnd is
upside down, the single point handler will always expect another event
than physically possible: For example, when the handler expects
TouchEnd and no more fingers are on the touch screen, the expected
event will always be preceded by a TouchBegin, which makes the
handler wait for another TouchEnd. Ultimately, the application can
no longer interpret incoming touch events, because they will be
ignored as unexpected.

This patch adds functionality to the handling of unexpected events.
If an unexpected TouchBegin event with exactly one event point is
observed, the current gesture is canceled and uses the stray event
as a new starting point.

An autotest is added in tst_QQuickTapHandler::touchMultiTap().

Fixes: QTBUG-108689
Fixes: QTBUG-113005
Pick-to: 6.5
Change-Id: I344e77e87acbaab8e4033606ca3f244427b94458
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>