qmlls: find usages of id's and properties
Implement find usages for id's and properties. Use the QQmlJSScope to do
typeresolution to avoid finding wrong usages, e.g. properties with the
same name but that are on a different QmlObject.

Implement find Usages to find usages of id's and properties. Use the type
resolution of the QQmlJSScope to find the correct usages in
fieldmemberexpressions, e.g. when multiple qml objects have a property
with the same name that should not be confused with each other.
To resolve id's to the correct qml objects, use the qqmljstyperesolver
which is inserted in the Dom during type construction.

During the DomConstruction, set semantic scopes inside of QmlComponents
and PropertyDefinitions. The latter one has to be done before the
endVisit call, so split
QQmlDomAstCreatorWithQQmlJSScope::setScopeInDom()
into setScopeInDomBeforeEndvisit() and setScopeInDomAfterEndvisit().

Task-number: QTBUG-113381
Task-number: QTBUG-92876
Task-number: QTBUG-111415
Change-Id: I4818bf8624fb969b3e9fc5f622403f582559addd
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>