QmlDom: implement the FieldMemberExpression
Implement expressions like "someId.propertyX" in QmlDom as a
BinaryExpression. Fix the BinaryExpression to have an enum as operator,
but only implement it for FieldMemberExpression and
ArrayMemberExpressions (as other operations are ignored by qmlls for now).

This is needed for qmlls to be able to find usages of QML object ids and
properties when using 'someId.propertyX' or 'someIf["propertyX"]' in JS code.

Also, replace one assert in qqmllsutils with a warning, as it crashes
unrelated tests. The checked condition is indeed not critical, in the
worst case some DomItem might not be able to be looked up from their text
position. See test in following commit.

Fixes: QTBUG-113381
Task-number: QTBUG-92876
Task-number: QTBUG-111415
Change-Id: I85c0703ec0b6e6f90962c4b3172ff084c3c2d606
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>