Remove the 'qml-devtools' feature
The meaning of the feature has changed and tools that depends on this
feature now need to be built unconditionally. So the feature should be
removed to avoid confusing users.

[ChangeLog][QtQml][Tools] The 'qml-devtools' feature is removed. All
tools that depend on this feature are mandatory and need to be build
unconditionally.

Fixes: QTBUG-113221
Change-Id: Ia026684703d847ce777d7a2916f04b3c90727695
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 41f36038fb2618561b2405efbdaf7ccd5bf3769d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>