Silence signed vs unsigned warning on 32 bit VS 2022 debug build
The warning C4018: '<': signed/unsigned mismatch appears in a Q_ASSERT
when checking a lookup index against the size of the container.

Fixed by changing from unsigned to signed type for the index. Since the
index is already implicitly converted to signed type when used, we can
use a signed type from the start. We rely on implicit conversion from
unsigned instead of static_cast to not hide other warnings in the
future.

Change-Id: I2b1983bdd40104e2c7135eec849a198ac074517c
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>