QmlDom: Add GenericScriptElement and CallExpression
Instead of implementing each scriptelement separately, implement all the
missing ones using a GenericScriptElement.
Also implement callexpressions with a GenericScriptElement.

Remove the empty destructor overrides of the scriptelement
implementations as they are not needed.

Write a test that really tests if the arguments of call-expressions are
correctly collected.
Do not make test_qmlformat fail on the call expressions qml file.

Task-number: QTBUG-92876
Change-Id: I8f17478cc3840ef83064d7201cb6d67d97225b5f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>