Change ConfigReader and templates according to the updated config file format
- Remove the now unnecessary configForImageUrl method of the
ConfigReader. The config can now be retrieved simply through
the "controls" property
- Change the QML templates to access the properties according
to the new format: eg.: config.background.leftOffset
- Refactor the QML templates a bit. Get rid of the ImageSelector
since that relies on an image always being available and will
find the image with the best score, eg.: in case of a control
in the checked & hovered state, if no image is specified for
this state, it will fallback to using the image for the checked
state. This is not what we want, as we are exporting images
for each of the states, and if there is no image for one,
that only means that that particular state doesn't use an
image, eg.: the background should be transparent in that
state

Task-number: QTBUG-111560
Change-Id: I7edf67a5e8b5ccb21566862d2390c90c66e5e30f
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>