TableView: implement SelectionMode
This patch will implement support for setting a
selectionMode. This API mirrors the selectionMode
API in QTableView.

Currently only SingleSelection, ContiguousSelection
and ExtendedSelection are implemented. The last mode,
MultiSelection, is postponed for now since its unclear
which use-case it supports, and also seems to overlap
too much with ExtendedSelection (if holding down Control
modifier). NoSelection is already supported by setting
selectionBehavior to TableView.SelectionDisabled.

[ChangeLog][Quick][TableView] You can now set a selectionMode
in TableView, to control if the user should be allowed to
select single or multiple cells.

Fixes: QTBUG-112836
Fixes: QTBUG-112835
Change-Id: I0f96d9fb68882aededc42d653c861cfd22e3d03a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>