QML: Maintain invariant between QObjectMethod members
If the methodCount is 0, the methods have to be nullptr. Otherwise they
have to point to the actual method(s). This is important for the method
resolution to work correctly.

In particular when cloning a method we have to check for 0.

Amends commit 17bd07cbc5b6cf54716e991765ab3088a710d7b3.

Pick-to: 6.5
Fixes: QTBUG-113484
Change-Id: Ic31d6e391c1d74a162820232f242a19379f5e4df
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>