StyleGenerator: Generate transition information for state changes
In order to be able to convert the transitions between the
state changes defined in Figma to QML animations, we need
to firstly export those transition settings in the generated
config.json file.

The way it works in Figma is: the transition is applied to
the component, ie.: the state variant, and contains
information about the state to transition to, duration of
transition, and easing type. When the transition is applied
to the component, this means that each of the component's
atoms will apply the same transition during the state change.
Since the transition in Figma is applied to the component as
a whole, and not to each atom, in the config file too this
information will be "per-state".

There is a limitation in the Figma REST API which only
allows us to retrieve information for a single transition
per component, no matter how many transitions are applied to
the component in Figma. So for now there will only be one
transition per state change, even though "transitions" is
a JSON array.

Task-number: QTBUG-111560
Change-Id: Icf1a2e10ad96d4b95165bdeb533efdabe56e3f8f
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>